Add git diff
This will be obvious to some, but "git diff" is the same as "diff -u". I did not
have "diff" installed because I have git. However I was not sure the output was
the same so I installed it to make sure. This change should clarify that for
people who do not have "diff" installed.